+++ tests/test_package_contract.py
@@ -97,7 +97,7 @@ class PackageContractTests(unittest.TestCase):
             {
                 "source": "url",
                 "url": "https://github.com/johnrucnapier-sketch/ACGM-for-Codex.git",
-                "ref": "v0.1.0-rc.2",
+                "ref": "v0.1.0-rc.3",
             },
         )
         self.assertEqual(